requirements.txt
streamlit
fastapi
pydantic
uvicorn
python-dotenv
langchain
langchain_google_genai
langserve